更新了部分优惠券

+++ goods/src/main/java/cn/edu/xmu/goods/model/vo/ActivityFinderVo.java
@@ -18,5 +18,5 @@ public class ActivityFinderVo {
     Integer pageSize;
 
     @ApiModelProperty(value = "时间线")
-    Integer timeline;
+    Byte timeline;
 }
